+++ Runtime/Schemas/ArenaSphereJson.cs
@@ -56,20 +56,20 @@ namespace ArenaUnity.Schemas
             return (Radius != defRadius);
         }
 
-        private static float defSegmentsHeight = 18f;
+        private static int defSegmentsHeight = 18;
         [JsonProperty(PropertyName = "segmentsHeight")]
         [Tooltip("segments height")]
-        public float SegmentsHeight = defSegmentsHeight;
+        public int SegmentsHeight = defSegmentsHeight;
         public bool ShouldSerializeSegmentsHeight()
         {
             if (_token != null && _token.SelectToken("segmentsHeight") != null) return true;
             return (SegmentsHeight != defSegmentsHeight);
         }
 
-        private static float defSegmentsWidth = 36f;
+        private static int defSegmentsWidth = 36;
         [JsonProperty(PropertyName = "segmentsWidth")]
         [Tooltip("segments width")]
-        public float SegmentsWidth = defSegmentsWidth;
+        public int SegmentsWidth = defSegmentsWidth;
         public bool ShouldSerializeSegmentsWidth()
         {
             if (_token != null && _token.SelectToken("segmentsWidth") != null) return true;
